requirements.txt
# Core Flask dependencies
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-CORS==6.0.0
Flask-RESTX==1.3.0
Flask-Limiter==3.8.0

# Database
PyMySQL==1.1.1
SQLAlchemy==2.0.36

# Caching
redis==5.2.0

# Environment and Configuration
python-dotenv==1.0.1
marshmallow==3.22.0

# HTTP Client for service communication
requests==2.32.4
httpx==0.28.1

# Message Broker
aio-pika==9.5.7
aiokafka==0.12.0

# System monitoring
psutil==6.0.0

# Utilities
uuid==1.30
python-dateutil==2.9.0

# Development dependencies
pytest==8.3.3
pytest-flask==1.3.0
pytest-cov==6.0.0
pytest-mock==3.14.0
black==24.8.0
flake8==7.1.1
isort==5.13.2

# Production
gunicorn==23.0.0
